requirements.txt
anyio==4.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
matrix-client==0.4.0
ollama==0.2.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
urllib3==1.26.18
